+++ ACP3/Modules/ACP3/Gallery/Installer/Migration.php
@@ -61,6 +61,9 @@ class Migration extends \ACP3\Core\Installer\AbstractMigration
             41 => [
                 "ALTER TABLE `{pre}gallery` ADD COLUMN `updated_at` DATETIME NOT NULL AFTER `end`;",
                 "UPDATE `{pre}gallery` SET `updated_at` = `start`;"
+            ],
+            42 => [
+                "INSERT INTO `{pre}acl_resources` (`id`, `module_id`, `area`, `controller`, `page`, `params`, `privilege_id`) VALUES('', '{moduleId}', 'admin', 'pictures', 'index', '', 3);",
             ]
         ];
     }