+++ ACP3/Modules/ACP3/System/Validation/AdminSettingsFormValidation.php
@@ -114,10 +114,10 @@ class AdminSettingsFormValidation extends Core\Validation\AbstractFormValidation
                 Core\Validation\ValidationRules\InArrayValidationRule::class,
                 [
                     'data' => $formData,
-                    'field' => 'site_subtitle_homepage_mode',
-                    'message' => $this->translator->t('system', 'select_site_subtitle_homepage_mode'),
+                    'field' => 'site_subtitle_mode',
+                    'message' => $this->translator->t('system', 'select_site_subtitle_mode'),
                     'extra' => [
-                        'haystack' => [0, 1]
+                        'haystack' => [1, 2, 3]
                     ]
                 ]
             )
@@ -125,10 +125,10 @@ class AdminSettingsFormValidation extends Core\Validation\AbstractFormValidation
                 Core\Validation\ValidationRules\InArrayValidationRule::class,
                 [
                     'data' => $formData,
-                    'field' => 'site_subtitle_mode',
-                    'message' => $this->translator->t('system', 'select_site_subtitle_mode'),
+                    'field' => 'site_subtitle_homepage_mode',
+                    'message' => $this->translator->t('system', 'select_site_subtitle_homepage_mode'),
                     'extra' => [
-                        'haystack' => [1, 2]
+                        'haystack' => [0, 1]
                     ]
                 ]
             );