+++ CHANGELOG.md
@@ -44,7 +44,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Updated the minimal required Symfony version to 3.3 and took advantage of some features of it
 - Updated the friendsofsymfony/http-cache package to version 2.*
 
-## [Unreleased] - 2017-xx-xx
+## [Unreleased] - 2017-xx-yy
+### Added
+- Added the new data grid column renderer `RouteColumnRenderer` which makes it possible specify and open routes in new tabs
+
+### Changed
+- Decoupled the menu item management from the articles module some more
+
+### Fixed
+- Fixed the bug that it was not possible to create new articles when the menu items module was installed and active, but there were no menus at all
+
+## [4.16.0] - 2017-10-24
+### Added
+- Added a new option which completely disables the site subtitle program logic 
+
+### Changed
+- extracted the gallery pictures data grid into a separate controller action
+
+## [4.15.0] - 2017-10-23
 ### Added
 - Added the schema.org `BreadcrumbList` markup to the breadcrumb template file 
 - Added the possibility to add a separate page title via the SEO form fields
@@ -458,7 +475,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - It should be possible again to save ACL roles
 - Silenced a possible PHP warning when trying to login with incorrect credentials 
 
-[Unreleased]: https://github.com/ACP3/cms/compare/v4.14.0...HEAD
+[Unreleased]: https://github.com/ACP3/cms/compare/v4.16.0...HEAD
+[4.16.0]: https://github.com/ACP3/cms/compare/v4.15.0...v4.16.0
+[4.15.0]: https://github.com/ACP3/cms/compare/v4.14.0...v4.15.0
 [4.14.0]: https://github.com/ACP3/cms/compare/v4.13.1...v4.14.0
 [4.13.1]: https://github.com/ACP3/cms/compare/v4.13.0...v4.13.1
 [4.13.0]: https://github.com/ACP3/cms/compare/v4.12.2...v4.13.0